Promote the shared no-other-transactions scan into the convertible checks file
Cancellation and conversion carried identical copies of the scan, differing
only in how each phrased the self-exemption. The shared check exempts the
transaction under validation by id, which is the invariant both copies
encoded.
