+++ ocf_validator/validators/convertible/tx_convertible_cancellation.ts
@@ -1,38 +1,5 @@
-import { defineValidator, type CheckObject } from "../checkKit";
-import { issuanceExists } from "./checks";
-
-const noOtherTransactions = {
-  id: "no-other-transactions",
-  severity: "error",
-  description:
-    "No other transaction dated on or before this transaction references its security_id, other than a convertible acceptance.",
-  run: (context, data: { id: string; security_id: string; date: string }) => {
-    const messages: string[] = [];
-    const { transactions } = context.ocfPackageContent;
-
-    // no-other-transactions emits one finding per transaction on this security_id
-    // dated on or before this cancellation, exempting the issuance, any convertible
-    // acceptance, and this cancellation itself. This scan keeps every transaction
-    // type in play, so it narrows by property presence rather than by discriminant:
-    // a transaction that carries no security_id can never reference this one.
-    transactions.forEach((ele) => {
-      if (
-        "security_id" in ele &&
-        ele.security_id === data.security_id &&
-        ele.date <= data.date &&
-        ele.object_type !== "TX_CONVERTIBLE_ISSUANCE" &&
-        ele.object_type !== "TX_CONVERTIBLE_ACCEPTANCE" &&
-        !(ele.object_type === "TX_CONVERTIBLE_CANCELLATION" && ele.id === data.id)
-      ) {
-        messages.push(
-          `Another transaction (${ele.id}) references the transaction's security_id.`,
-        );
-      }
-    });
-
-    return messages;
-  },
-} satisfies CheckObject;
+import { defineValidator } from "../checkKit";
+import { issuanceExists, noOtherTransactions } from "./checks";
 
 export const TX_CONVERTIBLE_CANCELLATION = defineValidator({
   transaction: "TX_CONVERTIBLE_CANCELLATION",
